Backport 65907ffdb155221f19ce8753ce4f13b5e096c64a
