+++ test/jdk/com/sun/security/auth/callback/TextCallbackHandler/Default.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2001, 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2001, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -23,31 +23,48 @@
 
 /*
  * @test
+ * @library /test/lib /java/security/testlibrary
  * @bug 4470717
  * @summary fix default handling and other misc
- * @run main/manual Default
+ * @run main/othervm Default
  */
 
 import com.sun.security.auth.callback.TextCallbackHandler;
+import jdk.test.lib.Asserts;
+
 import javax.security.auth.callback.*;
+import java.io.*;
 
 public class Default {
-   public static void main(String args[]) throws Exception {
-        TextCallbackHandler h = new TextCallbackHandler();
-        NameCallback nc = new NameCallback("Name: ", "charlie");
-        ConfirmationCallback cc = new ConfirmationCallback
-                        ("Correct?",
-                        ConfirmationCallback.INFORMATION,
-                        ConfirmationCallback.YES_NO_OPTION,
-                        ConfirmationCallback.NO);
+    public static void main(String args[]) throws Exception {
+        InputStream in = System.in;
+        PrintStream err = System.err;
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        final String defaultName = "charlie";
+        final String simulatedInput = "-1\n-1\n";
+        HumanInputStream humanInputStream = new HumanInputStream(simulatedInput);
 
-        Callback[] callbacks = { nc, cc };
-        h.handle(callbacks);
+        try (PrintStream prints = new PrintStream(baos)) {
+            System.setIn(humanInputStream);
+            System.setErr(prints);
+            NameCallback nameCallback = new NameCallback("Name: ", defaultName);
+            ConfirmationCallback confirmationCallback = new ConfirmationCallback(
+                    "Correct?",
+                    ConfirmationCallback.INFORMATION,
+                    ConfirmationCallback.YES_NO_OPTION,
+                    ConfirmationCallback.NO);
+            new TextCallbackHandler().handle(new Callback[]{nameCallback, confirmationCallback});
 
-        if (cc.getSelectedIndex() == ConfirmationCallback.YES) {
-            System.out.println("yes");
-        } else {
-            System.out.println("no");
+            Asserts.assertEquals(nameCallback.getDefaultName(), defaultName);
+            Asserts.assertEquals(confirmationCallback.getSelectedIndex(), ConfirmationCallback.NO);
+
+        } finally {
+            System.setIn(in);
+            System.setErr(err);
         }
-   }
+
+        // check that the default name and confirmation were visible in the output
+        Asserts.assertTrue(baos.toString().contains(String.format("Name:  [%s]", defaultName)));
+        Asserts.assertTrue(baos.toString().contains("1. No [default]"));
+    }
 }
